+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## Next
+
+- refactor: enchance dns resolver to update upstreams.
+
 ## 2024-12-22 0.1.0
 
 - all features are implemented and allow for previewing